Fix compilation errors for Kotlin/Android build
Resolves 9 compilation errors identified during Android Studio build:

1. AudioEncoder.kt:66 - Fix integer literal overflow
   - Cast 0xFFFFFFFE to Int explicitly to match expected type

2. DeepLTranslationService.kt:39-42 - Remove unsupported timeout DSL
   - Ktor HttpTimeout plugin not configured in HttpClient
   - Removed inline timeout block (can be configured globally later)

3. TranslationBubble.kt:33 - Fix alignment type mismatch
   - Changed from Alignment.CenterEnd/CenterStart to Alignment.End/Start
   - Column requires Alignment.Horizontal, not general Alignment

4. RealtimeTranslationService.kt:49-51 - Fix WebSocket connection syntax
   - Changed from url() + header() DSL to webSocketSession(urlString, request)
   - Correct API for Ktor 2.3.6 WebSocket client

5. RealtimeTranslationService.kt:73 - Fix launch() scope issue
   - Changed flow {} to channelFlow {} to provide CoroutineScope
   - channelFlow provides ProducerScope with launch() function

6. RealtimeTranslationService.kt:229 - Fix WebSocket close
   - Changed cancel() to close() - WebSocketSession is Closeable
   - Added catch block for close exceptions

7. app/build.gradle.kts - Add ktor-client-resources dependency
   - Ensures all Ktor client features are available

All compilation errors resolved. Ready for build.

+++ app/src/main/java/com/dragoncloser/app/data/audio/AudioEncoder.kt
@@ -63,7 +63,7 @@ class AudioEncoder @Inject constructor() {
         if (sourceSampleRate == targetSampleRate) return pcmData
 
         val ratio = sourceSampleRate.toDouble() / targetSampleRate
-        val outputSize = (pcmData.size / ratio).toInt() and 0xFFFFFFFE // Ensure even number
+        val outputSize = (pcmData.size / ratio).toInt() and 0xFFFFFFFE.toInt() // Ensure even number
 
         val result = ByteArray(outputSize)
         var outputIndex = 0
